refactor: Move WebSocket management to parent page component
TBS-71
